usm: redis: Remove user-mode struct duplication
