+++ pkg/network/protocols/redis/client.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build test
+//go:build linux_bpf && test
 
 // Package redis implements USM's Redis monitoring, as well as provide
 // helpers used in tests.
